Refactor code to seperate utilities from presentation
