+++ polaris-react/src/components/Banner/components/PolarisSummerEditions2023Banner/PolarisSummerEditions2023Banner.tsx
@@ -1,0 +1,140 @@
+import React, {useContext} from 'react';
+
+import {Text} from '../../../Text';
+import {VerticalStack} from '../../../VerticalStack';
+import {HorizontalStack} from '../../../HorizontalStack';
+import {useBreakpoints} from '../../../../utilities/breakpoints';
+import {WithinContentContext} from '../../../../utilities/within-content-context';
+import {Box} from '../../../Box';
+import {Button} from '../../../Button';
+import {ButtonGroup} from '../../../ButtonGroup';
+import type {BannerProps} from '../../Banner';
+import {Icon} from '../../../Icon';
+
+import {useBannerColors} from './utilities';
+
+export function PolarisSummerEditions2023Banner({
+  status = 'info',
+  icon,
+  hideIcon,
+  onDismiss,
+  action,
+  secondaryAction,
+  title,
+  children,
+}: BannerProps) {
+  const withinContentContainer = useContext(WithinContentContext);
+  const {smDown} = useBreakpoints();
+  const {iconRGBA, backgroundColor, textColor, statusIcon, closeIcon} =
+    useBannerColors(status);
+
+  const bannerIcon = hideIcon ? null : (
+    <Box paddingInlineStart="05">
+      {icon ? (
+        <span style={{fill: iconRGBA}}>
+          <Icon source={icon} />
+        </span>
+      ) : (
+        <Icon source={statusIcon} />
+      )}
+    </Box>
+  );
+
+  const dismissButton = onDismiss ? (
+    <Button
+      plain
+      icon={closeIcon}
+      onClick={onDismiss}
+      // se23: hard coded string copied from Banner, TODO make i18n key
+      accessibilityLabel="Dismiss notification"
+    />
+  ) : null;
+
+  const actionButtons =
+    action || secondaryAction ? (
+      <ButtonGroup>
+        {action && (
+          <Button onClick={action.onAction} {...action}>
+            {action.content}
+          </Button>
+        )}
+        {secondaryAction && (
+          <Button onClick={secondaryAction.onAction} {...secondaryAction}>
+            {secondaryAction.content}
+          </Button>
+        )}
+      </ButtonGroup>
+    ) : null;
+
+  const bannerTitle = title ? (
+    <Text as="h2" variant="headingSm" breakWord>
+      {title}
+    </Text>
+  ) : null;
+
+  const hasContent = children || action || secondaryAction;
+
+  return withinContentContainer ? (
+    <Box
+      width="100%"
+      background={backgroundColor}
+      padding="2"
+      borderRadius="2"
+      color={textColor}
+    >
+      <HorizontalStack align="space-between" blockAlign="start" wrap={false}>
+        <Box paddingInlineEnd={onDismiss ? '6' : undefined}>
+          <HorizontalStack gap="2" wrap={false}>
+            {bannerIcon}
+            <VerticalStack gap="2">
+              <VerticalStack gap="05">
+                {bannerTitle}
+                {children}
+              </VerticalStack>
+              {actionButtons}
+            </VerticalStack>
+          </HorizontalStack>
+        </Box>
+        <Box position="absolute" insetBlockStart="3" insetInlineEnd="3">
+          {dismissButton}
+        </Box>
+      </HorizontalStack>
+    </Box>
+  ) : (
+    <Box width="100%">
+      <VerticalStack align="space-between">
+        <Box
+          background={backgroundColor}
+          color={textColor}
+          borderRadiusStartStart={smDown ? undefined : '2'}
+          borderRadiusStartEnd={smDown ? undefined : '2'}
+          borderRadiusEndStart={hasContent || smDown ? undefined : '2'}
+          borderRadiusEndEnd={hasContent || smDown ? undefined : '2'}
+          padding={{xs: '2', sm: '3'}}
+          paddingInlineEnd={{xs: '3', sm: '4'}}
+        >
+          <HorizontalStack
+            align="space-between"
+            blockAlign="center"
+            gap="2"
+            wrap={false}
+          >
+            <HorizontalStack gap="2" wrap={false}>
+              {bannerIcon}
+              {bannerTitle}
+            </HorizontalStack>
+            {dismissButton}
+          </HorizontalStack>
+        </Box>
+        {hasContent && (
+          <Box padding={{xs: '3', sm: '4'}} paddingBlockStart="3">
+            <VerticalStack gap="2">
+              {children}
+              {actionButtons}
+            </VerticalStack>
+          </Box>
+        )}
+      </VerticalStack>
+    </Box>
+  );
+}
