+++ polaris-react/src/components/Banner/components/index.ts
@@ -1,0 +1,1 @@
+export {PolarisSummerEditions2023Banner} from './PolarisSummerEditions2023Banner';
