+++ polaris-react/src/components/Button/Button.scss
@@ -28,38 +28,9 @@
   // stylelint-disable-next-line -- generated by polaris-migrator DO NOT COPY
   @include button-base;
 
-<<<<<<< HEAD
   &.disabled {
     // stylelint-disable-next-line -- generated by polaris-migrator DO NOT COPY
     @include base-button-disabled;
-=======
-  #{$se23} & {
-    --p-depth-secondary-outset-2: inset 0px -1px 0px #b5b5b5,
-      inset -1px 0px 0px #e3e3e3, inset 1px 0px 0px #e3e3e3,
-      inset 0px 1px 0px #e3e3e3;
-    --pc-button-box-shadow: var(--p-depth-secondary-outset-2);
-    --pc-button-border-disabled-border-box-shadow: 0px 0px 0px 1px #ebebeb;
-    box-shadow: var(--pc-button-box-shadow);
-
-    transition: box-shadow 75ms cubic-bezier(0.19, 0.91, 0.38, 1);
-
-    &:active {
-      --pc-button-box-shadow: inset -1px 0px 1px rgba(0, 0, 0, 0.12),
-        inset 1px 0px 1px rgba(0, 0, 0, 0.12),
-        inset 0px 2px 1px rgba(0, 0, 0, 0.2);
-      .Content {
-        transform: translateY(1px);
-      }
-    }
-  }
-
-  &.disabled {
-    // stylelint-disable-next-line -- generated by polaris-migrator DO NOT COPY
-    @include base-button-disabled;
-    #{$se23} & {
-      box-shadow: var(--pc-button-border-disabled-border-box-shadow);
-    }
->>>>>>> 5e48bd9e7 (Add summer editions updates to the Banner component)
   }
 
   &.connectedDisclosure {
@@ -236,33 +207,6 @@
 }
 
 .primary {
-<<<<<<< HEAD
-=======
-  #{$se23} & {
-    &::before {
-      content: '';
-      position: absolute;
-      left: 2px;
-      right: 2px;
-      top: 1.5px;
-      bottom: 3.5px;
-
-      /* shine */
-      background: linear-gradient(
-        180deg,
-        rgba(255, 255, 255, 0.2) 0%,
-        rgba(255, 255, 255, 0) 36.84%
-      );
-      box-shadow: inset 0px 1px 0px rgba(255, 255, 255, 0.55);
-      border-radius: 6px;
-      opacity: 1;
-      transition: opacity 75ms cubic-bezier(0.19, 0.91, 0.38, 1);
-    }
-  }
-}
-
-.primary {
->>>>>>> 5e48bd9e7 (Add summer editions updates to the Banner component)
   // stylelint-disable -- Polaris component custom properties
   --pc-button-color: var(--p-color-bg-primary);
   --pc-button-text: var(--p-color-text-on-color);
@@ -270,65 +214,6 @@
   --pc-button-color-active: var(--p-color-bg-primary-active);
   --pc-button-color-depressed: var(--p-color-bg-primary-active);
   // stylelint-enable
-<<<<<<< HEAD
-=======
-
-  #{$se23} & {
-    &:active {
-      --pc-button-box-shadow: inset -1px 0px 1px rgba(0, 0, 0, 0.2),
-        inset 1px 0px 1px rgba(0, 0, 0, 0.2),
-        inset 0px 2px 1px rgba(0, 0, 0, 0.6);
-      &::before {
-        opacity: 0;
-      }
-    }
-    &.disabled {
-      color: var(--pc-button-text);
-
-      &::before {
-        display: none;
-      }
-
-      svg {
-        fill: var(--pc-button-text);
-      }
-    }
-
-    &.destructive {
-      // stylelint-disable -- Polaris component custom properties
-      --pc-button-color: var(--p-color-bg-critical-strong);
-      --pc-button-text: var(--p-color-text-on-color);
-      --pc-button-color-hover: var(--p-color-bg-critical-strong-hover);
-      --pc-button-color-active: var(--p-color-bg-critical-strong-active);
-      --pc-button-color-depressed: var(--p-color-bg-critical-strong-active);
-      // stylelint-enable
-      svg {
-        fill: var(--p-color-icon-on-color);
-      }
-
-      &.disabled {
-        background: var(--p-color-bg-disabled);
-
-        &::before {
-          display: none;
-        }
-        svg {
-          fill: var(--pc-button-text);
-        }
-      }
-    }
-
-    &.success {
-      // stylelint-disable -- Polaris component custom properties
-      --pc-button-color: var(--p-color-bg-success-strong);
-      --pc-button-text: var(--p-color-text-on-color);
-      --pc-button-color-hover: var(--p-color-bg-success-strong-hover);
-      --pc-button-color-active: var(--p-color-bg-success-strong-active);
-      --pc-button-color-depressed: var(--p-color-bg-success-strong-active);
-      // stylelint-enable
-    }
-  }
->>>>>>> 5e48bd9e7 (Add summer editions updates to the Banner component)
 
   svg {
     fill: var(--p-color-icon-on-color);
@@ -492,26 +377,6 @@
 // in this case.
 // stylelint-disable selector-max-specificity, selector-max-class -- generated by polaris-migrator DO NOT COPY
 .plain {
-<<<<<<< HEAD
-=======
-  #{$se23} & {
-    // stylelint-disable -- Polaris component custom properties
-    --p-color-text-interactive: --p-color-text;
-
-    &:hover {
-      background: rgba(26, 26, 26, 0.05);
-    }
-
-    &:focus-visible {
-      background: none;
-    }
-
-    &:active {
-      background: rgba(26, 26, 26, 0.07);
-    }
-  }
-
->>>>>>> 5e48bd9e7 (Add summer editions updates to the Banner component)
   // stylelint-disable-next-line -- polaris/conventions/polaris/custom-property-allowed-list
   margin: calc(-1 * var(--pc-button-vertical-padding))
     calc(-1 * var(--p-space-2));
@@ -641,9 +506,6 @@
 
   &.iconOnly {
     margin: calc(-1 * var(--p-space-2));
-    #{$se23} & {
-      padding: 0;
-    }
 
     svg {
       fill: var(--p-color-icon);
