+++ polaris-react/src/components/Text/Text.scss
@@ -56,10 +56,6 @@
   color: var(--p-color-text-inverse);
 }
 
-.text-on-color {
-  color: var(--p-color-text-on-color);
-}
-
 .headingXs {
   font-size: var(--p-font-size-75);
   line-height: var(--p-font-line-height-1);
